feat: move interest calculation functions to a new module and update imports
